requirements.txt
argcomplete==3.6.2
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
bcc==0.1.10
bcrypt==4.3.0
blinker==1.9.0
Django==5.2
django-filter==25.1
djangorestframework==3.16.1
dotenv==0.9.9
numpy==2.2.4
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
traitlets==5.14.3
traittypes==0.2.1
